requirements.txt
git+https://github.com/huggingface/transformers.git@main
sentence_transformers
torch
accelerate
bitsandbytes
scipy
Pillow
scikit-learn
numpy
matplotlib
tqdm

# ClusterEnsembles
# ClustPy